+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @couch-kit/cli
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @couch-kit/core@0.5.0
+
 ## 0.2.2
 
 ### Patch Changes